requirements.txt
Flask==1.1.2
setuptools==40.8.0
google-cloud-ndb
googleapis_common_protos
